Move AWS Glue from contrib to providers
