+++ tests/contrib/hooks/test_aws_glue_catalog_hook.py
@@ -20,7 +20,7 @@ import unittest
 
 import boto3
 
-from airflow.contrib.hooks.aws_glue_catalog_hook import AwsGlueCatalogHook
+from airflow.providers.amazon.aws.hooks.glue import AwsGlueCatalogHook
 from tests.compat import mock
 
 try:
