+++ tests/contrib/sensors/test_aws_glue_catalog_partition_sensor.py
@@ -19,8 +19,8 @@
 
 import unittest
 
-from airflow.contrib.hooks.aws_glue_catalog_hook import AwsGlueCatalogHook
-from airflow.contrib.sensors.aws_glue_catalog_partition_sensor import AwsGlueCatalogPartitionSensor
+from airflow.providers.amazon.aws.hooks.glue import AwsGlueCatalogHook
+from airflow.providers.amazon.aws.sensors.glue import AwsGlueCatalogPartitionSensor
 from tests.compat import mock
 
 try:
